+++ src/pgfimage.cpp
@@ -139,7 +139,7 @@ namespace Exiv2 {
 
         DataBuf imgData(size);
         std::memset(imgData.pData_, 0x0, imgData.size_);
-        long bufRead = io_->read(imgData.pData_, imgData.size_);
+        size_t bufRead = io_->read(imgData.pData_, imgData.size_);
         if (io_->error()) throw Error(kerFailedToReadImageData);
         if (bufRead != imgData.size_) throw Error(kerInputDataReadFailed);
 
@@ -238,7 +238,7 @@ namespace Exiv2 {
         // Copy the rest of PGF image data.
 
         DataBuf buf(4096);
-        long readSize = 0;
+        size_t readSize = 0;
         while ((readSize=io_->read(buf.pData_, buf.size_)))
         {
             if (outIo.write(buf.pData_, readSize) != readSize) throw Error(kerImageWriteFailed);
@@ -266,7 +266,7 @@ namespace Exiv2 {
     uint32_t PgfImage::readPgfHeaderSize(BasicIo& iIo)
     {
         DataBuf buffer(4);
-        long bufRead = iIo.read(buffer.pData_, buffer.size_);
+        size_t bufRead = iIo.read(buffer.pData_, buffer.size_);
         if (iIo.error()) throw Error(kerFailedToReadImageData);
         if (bufRead != buffer.size_) throw Error(kerInputDataReadFailed);
 
@@ -283,7 +283,7 @@ namespace Exiv2 {
     DataBuf PgfImage::readPgfHeaderStructure(BasicIo& iIo, int& width, int& height)
     {
         DataBuf header(16);
-        long bufRead = iIo.read(header.pData_, header.size_);
+        size_t bufRead = iIo.read(header.pData_, header.size_);
         if (iIo.error()) throw Error(kerFailedToReadImageData);
         if (bufRead != header.size_) throw Error(kerInputDataReadFailed);
